+++ include/ppMethod.h
@@ -4,10 +4,10 @@
 #include "vec3.h"
 
 std::string ppMethod(std::vector<Vec3>& state,
-                     std::vector<double>& masses,
+                     std::vector<float>& masses,
                      const int simLength,
-                     const double stepSize,
-                     const double G,
+                     const float stepSize,
+                     const float G,
                      const char* positionsPath = "output.txt",
                      const char* energyPath = "energy.txt",
                      const char* momentumPath = "momentum.txt");
