+++ include/unitConversions.h
@@ -2,22 +2,22 @@
 #include <vector>
 #include "vec3.h"
 
-Vec3 positionToCodeUntits(const Vec3& pos, double H);
+Vec3 positionToCodeUntits(const Vec3& pos, float H);
 
-Vec3 velocityToCodeUntits(const Vec3& v, double H, double DT);
+Vec3 velocityToCodeUntits(const Vec3& v, float H, float DT);
 
-Vec3 accelerationToCodeUnits(const Vec3& a, double H, double DT);
+Vec3 accelerationToCodeUnits(const Vec3& a, float H, float DT);
 
-double densityToCodeUnits(double density, double DT, double G);
+float densityToCodeUnits(float density, float DT, float G);
 
-void stateToCodeUnits(std::vector<Vec3>& state, double H, double DT);
+void stateToCodeUnits(std::vector<Vec3>& state, float H, float DT);
 
-void velocitiesToCodeUnits(std::vector<Vec3>& velocities, double H, double DT);
+void velocitiesToCodeUnits(std::vector<Vec3>& velocities, float H, float DT);
 
-Vec3 positionToOriginalUnits(const Vec3& pos, double H);
+Vec3 positionToOriginalUnits(const Vec3& pos, float H);
 
-Vec3 velocityToOriginalUnits(const Vec3& v, double H, double DT);
+Vec3 velocityToOriginalUnits(const Vec3& v, float H, float DT);
 
-void stateToOriginalUnits(std::vector<Vec3>& state, double H, double DT);
+void stateToOriginalUnits(std::vector<Vec3>& state, float H, float DT);
 
-void velocitiesToOriginalUnits(std::vector<Vec3>& velocities, double H, double DT);
+void velocitiesToOriginalUnits(std::vector<Vec3>& velocities, float H, float DT);
